chore: cleanup bootc.ts
The original template had a bootc.ts that we didn't use. None of the
other extensions have this pattern, so we've decided to just clean it
up for now. Tests merged in and extra files deleted.

Fixes #55.
