--- src/bootc.ts
@@ -1,35 +1,0 @@
-/**********************************************************************
- * Copyright (C) 2024 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ***********************************************************************/
-
-import type { ExtensionContext } from '@podman-desktop/api';
-
-export class BootC {
-  readonly #extensionContext: ExtensionContext;
-
-  constructor(readonly extensionContext: ExtensionContext) {
-    this.#extensionContext = extensionContext;
-  }
-
-  public async activate(): Promise<void> {
-    console.log('starting boot-c extension');
-  }
-
-  public async deactivate(): Promise<void> {
-    console.log('stopping boot-c extension');
-  }
-}
